requirements.txt
fastapi
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
llama-index>=0.10.0
llama-index-readers-web>=0.1.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv
llama-index-llms-anthropic
llama-index-embeddings-openai
anthropic>=0.18.0
beautifulsoup4
llama-index 
llama-index-llms-groq